Add support for avoiding some race conditions with Meteor Actions
When executing the callback after a successfully returning meteor action,
first we call the `returned` callback, and then reset the `pending` flag.

This normally works OK, but there are some situations that we might
want to do something inside the returned callback (for example,
firing another action), which is forbidden by the pending state of
the current action.

Therefore, now we provide a way for clearing the pending flag sooner,
by providing a function to the returned method to do so, if desired.