Fix incorrect use of Ant radio groups
It turns out that when we define elements inside radio groups,
we are not supposed to pass in a "checked" property to these elements.

Instead, we should pass in "value" to the group itself, and the it will
calculate the checked state of the elements automatically.

We missed this earlier because the two approaches produce almost identical
behavior, expect in situations when the state changes quickly.
In those cases, specifying checked directly can lead to inconsistent
results.