NavTabBar: use anchors on tabs + finish accidental leave protection
This change introduces Anchors for the tabs. This means that
we can support all the formatting / styling options that are
supported by the other nav menu.

Also, now we are always using the `trySelectSubState()` method upon
clicking, which supports not navigating away accidentally.

Also, fix the problem with the root states, by always using the
`menuKey` instead of `key`.