Eliminate duplicate change interceptors
... which has sneaked into the system when we started to
dynamically re-create the state space handlers.

Now each change interceptor must be able to produce a unique ID
for the component that it is representing, and newer iterations
of the same interceptor replace the old one's registration, instead
of being added at the end of the list.

This make sure that we always have only one interceptor from
each kind, and thus eliminates the duplicated questions.