Fix `mountAll` option
When using the `mountAll` option with a menu, rendering of "fallback"
content was broken. (Ie. the thing that is shown when a non-existent
item is requested from the menu.)

Also, we were rendering _all_ the sub-states, not only those that were
currently available, considering the conditions.

These problems have been fixed.