Make it possible to override the help on form items
When we have a BoolAnt, and we put it into a form via BoolFormAnt,
then the label is displayed twice: before and after the checkbox.

Since we usually don't want that, we have already set up FormItemAnt
so that the "label" prop can be overridden, so that we are not forced
to display the label coming directly for the operation.

The current change extends this support to the "help" field of the operation.
The help is displayed in a popup when the user hovers above the small (?) besides
the label; basically it's an extension of the label. With this change, it can be
overridden the same way as it works with the label.