Revert to the browser's native title feature, instead of using a custom tooltip.
We used this tooltip for display various help messages.
But in some situations this doesn't work, so in those cases we used the browser's title attribute.
In the end we had a mixture of the two.

It seemed to be OK, but it turns out, that it some browsers the two thing looks quite differently,
which is confusing. So this PR deprecates the tooltip-based feature, and stops using it in widgets.

I haven't covered the Semantic UI widgets, because we don't really maintain them anyway.