Added `ShallowStructObservable`
A helper to do to combine `observable.shallow` with `observable.struct`.
Useful when `observable.struct` causes too many reactions and
`observable` would make all children observable. It is used when new
versions of a object come form a rest call or a meteor subscription and\
you are interested in changes at the top level.