Update Navigation interfaces + fix a bug
This is a breaking API change. All application code needs to be (slightly) updated.

The main change is that whenever we are specifying a state-space, now we can also specify
some central piece of meta-data along with the list of sub-states.

This is useful for both debugging and other, application-specific purposes.

Furthermore, there was a bug which caused the navigation system to behave in an inconsistent way
in some cases when the definition of the menu is changing on the fly (not only our position).
That has been fixed.